requirements.txt
llama-index-core==0.12.5
llama-index-readers-file==0.4.1
llama-parse==0.5.17
llama_cpp_python==0.3.5